Features/react native setup (#7)
* chore: add PR and issue templates (#1)

* docs: add CONTRIBUTING guide (#2)

* chore: add PR and issue templates

* docs: added contribution guidelines

* updated readme (#5)

* Set up react native

* added realtime button

---------
