+++ FES-app/FES/app/RealTimeData.tsx
@@ -1,0 +1,2 @@
+import RealTimeData from '../screens/RealTimeData';
+export default RealTimeData;
