+++ FES-app/FES/screens/RealTimeData.tsx
@@ -1,0 +1,8 @@
+import { ThemedText } from '@/components/themed-text';
+import React from 'react';
+
+export default function RealTimeData() {
+    return (
+        <ThemedText>Real-Time Data Screen</ThemedText>
+    );
+}
